Change server ip-address. Add GET_hub func. Code refactoring.
Change server ip-address and add new endpoint for GET request.
Add GET_hub function that send GET request and parses JSON
reply to extract the data aboud compressor and switch the pin.

Move some code, add debug output.
Add a timer to querry the server every 500 ms.

Configure pin 22 as an output and switch it if compressor is on.